refactor : 코치 상세 정보 조회 리팩토링

+++ src/main/java/site/coach_coach/coach_coach_server/coach/dto/CoachDetailDto.java
@@ -1,6 +1,5 @@
 package site.coach_coach.coach_coach_server.coach.dto;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import jakarta.validation.constraints.NotNull;
@@ -39,7 +38,7 @@ public record CoachDetailDto(
 	boolean isContacted,
 	boolean isMatched,
 	int countOfReviews,
-	BigDecimal reviewRating,
+	double reviewRating,
 	boolean isLiked,
 	int countOfLikes
 ) {
